+++ docs/source/changelog.md
@@ -11,6 +11,7 @@ Release notes for the `space_packet_parser` library
 
 - Add support for filtering packets in `create_dataset`
 - BUGFIX: Handle optional secondary headers with CCSDS continuation packets
+- Add warnings if there are leftover bytes from a CCSDS generator
 
 ### v6.0.0 (released)